Change float64 to float32.

+++ code/qa_model.py
@@ -72,7 +72,7 @@ class Encoder(object):
                                                                  sequence_length=seq_len,
                                                                  initial_state_fw=state_fw,
                                                                  initial_state_bw=state_bw,
-                                                                 dtype=tf.float64)
+                                                                 dtype=tf.float32)
 
         # Concatenate two end hidden vectors for the final encoded
         # representation of inputs
@@ -85,6 +85,7 @@ class Encoder(object):
         concat_final_state = tf.concat(1, [final_fw_m_state, final_bw_m_state])
         logging.debug('Shape of concatenated BiRNN final hiden state is %s' % str(concat_final_state))
         return concat_hidden_states, concat_final_state, final_state
+
 
 class QASystem(object):
     def __init__(self, encoder, *args):
@@ -120,7 +121,7 @@ class QASystem(object):
 
         self.start_labels_placeholder=tf.placeholder(tf.int64,(None,))
         self.end_labels_placeholder=tf.placeholder(tf.int64,(None,))
-        self.mask_placeholder = tf.placeholder(tf.float64, (None, self.config.max_context_length))
+        self.mask_placeholder = tf.placeholder(tf.float32, (None, self.config.max_context_length))
 
         # ==== assemble pieces ====
         with tf.variable_scope("qa", initializer=tf.uniform_unit_scaling_initializer(1.0)):
@@ -188,8 +189,8 @@ class QASystem(object):
             zero_init = tf.constant_initializer(0)
             
             if simple:
-                W_s = tf.get_variable('Ws_s', shape=(2*d, d), initializer=xavier_init, dtype=tf.float64)
-                b_s = tf.get_variable('bs_s', shape=(d, ), initializer=xavier_init, dtype=tf.float64)
+                W_s = tf.get_variable('Ws_s', shape=(2*d, d), initializer=xavier_init, dtype=tf.float32)
+                b_s = tf.get_variable('bs_s', shape=(d, ), initializer=xavier_init, dtype=tf.float32)
 
                 QT = tf.transpose(Q, [0,2,1])
                 A = tf.nn.softmax(_batch_mat_mul(h, QT))
@@ -208,7 +209,7 @@ class QASystem(object):
                 embed_dim = d
 
             else:
-                w_s = tf.get_variable('w_s', shape=(3*d, ), initializer=xavier_init, dtype=tf.float64)
+                w_s = tf.get_variable('w_s', shape=(3*d, ), initializer=xavier_init, dtype=tf.float32)
 
                 h_aug = tf.tile(tf.expand_dims(h, 2), [1, 1, JQ, 1]) # [?, JP, JQ, d]
                 u_aug = tf.tile(tf.expand_dims(u, 1), [1, JP, 1, 1]) # [?. JP, JQ, d]
@@ -286,10 +287,10 @@ class QASystem(object):
         zero_init = tf.constant_initializer(0)
 
         with tf.variable_scope("simple_decoder"):
-            Wp_s = tf.get_variable('Wp_s', shape=(d, ), initializer=xavier_init, dtype=tf.float64)
-            Wp_e = tf.get_variable('Wp_e', shape=(d, ), initializer=xavier_init, dtype=tf.float64)
-            b_s  = tf.get_variable('b_s', shape=(), initializer=zero_init, dtype=tf.float64)
-            b_e  = tf.get_variable('b_e', shape=(), initializer=zero_init, dtype=tf.float64)
+            Wp_s = tf.get_variable('Wp_s', shape=(d, ), initializer=xavier_init, dtype=tf.float32)
+            Wp_e = tf.get_variable('Wp_e', shape=(d, ), initializer=xavier_init, dtype=tf.float32)
+            b_s  = tf.get_variable('b_s', shape=(), initializer=zero_init, dtype=tf.float32)
+            b_e  = tf.get_variable('b_e', shape=(), initializer=zero_init, dtype=tf.float32)
 
             with tf.variable_scope('answer_start'):
                 a_s  = tf.reshape(tf.matmul(tf.reshape(H_in, [-1, d]), tf.expand_dims(Wp_s, 1)), [-1, con_len]) + b_s
@@ -311,13 +312,13 @@ class QASystem(object):
 
         def lstm_decoder_cell(H_in, d):
             """Helper function to create a lstm decoder."""
-            Wp = tf.get_variable('Wp', shape=(d, ), initializer=xavier_init, dtype=tf.float64)
+            Wp = tf.get_variable('Wp', shape=(d, ), initializer=xavier_init, dtype=tf.float32)
             # implement a lstm with H_in as input and H_out where H_out will be used for the end position prediction
             cell = tf.nn.rnn_cell.LSTMCell(num_units=d, state_is_tuple=True)
             H_out, _ = tf.nn.dynamic_rnn(cell=cell,
                                         inputs=H_in,
                                         sequence_length=self.context_length_placeholder,
-                                        dtype=tf.float64)
+                                        dtype=tf.float32)
             a  = tf.reshape(tf.matmul(tf.reshape(H_out, [-1, d]), tf.expand_dims(Wp, 1)), [-1, con_len]) #  + b_s
             return a
 
@@ -384,7 +385,7 @@ class QASystem(object):
         """
         with vs.variable_scope("embeddings"):
             embedding_tensor = tf.Variable(self.pretrained_embeddings, trainable=False)
-            # embedding_tensor = tf.cast(self.pretrained_embeddings, tf.float64)
+            # embedding_tensor = tf.cast(self.pretrained_embeddings, tf.float32)
             question_embedding_lookup = tf.nn.embedding_lookup(embedding_tensor, self.question_placeholder)
             context_embedding_lookup = tf.nn.embedding_lookup(embedding_tensor, self.context_placeholder)
             question_embeddings = tf.reshape(question_embedding_lookup, [-1, self.config.max_question_length, self.config.embedding_size * self.config.n_features])
@@ -724,7 +725,7 @@ class QASystemMatchLSTM(QASystem):
 
         self.start_labels_placeholder=tf.placeholder(tf.int64,(None,))
         self.end_labels_placeholder=tf.placeholder(tf.int64,(None,))
-        self.mask_placeholder = tf.placeholder(tf.float64, (None, self.config.max_context_length))
+        self.mask_placeholder = tf.placeholder(tf.float32, (None, self.config.max_context_length))
 
         # ==== assemble pieces ====
         with tf.variable_scope("qa", initializer=tf.uniform_unit_scaling_initializer(1.0)):
@@ -775,7 +776,7 @@ class QASystemMatchLSTM(QASystem):
             H_p, _ = tf.nn.dynamic_rnn(cell=cell_p,
                                        inputs=passage,
                                        sequence_length=p_len,
-                                       dtype=tf.float64)
+                                       dtype=tf.float32)
 
         # LSTM Preprocessing Layer for question
         with tf.variable_scope('q'):
@@ -783,7 +784,7 @@ class QASystemMatchLSTM(QASystem):
             H_q, _ = tf.nn.dynamic_rnn(cell=cell_q,
                                        inputs=question,
                                        sequence_length=q_len,
-                                       dtype=tf.float64)
+                                       dtype=tf.float32)
 
         return H_p, H_q
 
@@ -794,12 +795,12 @@ class QASystemMatchLSTM(QASystem):
         max_question_length = self.config.max_question_length
       
         with tf.variable_scope('match_LSTM'):
-            W_q = tf.get_variable('W_q', shape=(self.config.state_size, self.config.state_size), initializer=xavier_init, dtype=tf.float64)
-            W_p = tf.get_variable('W_p', shape=(self.config.state_size, self.config.state_size), initializer=xavier_init, dtype=tf.float64)
-            W_r = tf.get_variable('W_r', shape=(self.config.state_size, self.config.state_size), initializer=xavier_init, dtype=tf.float64)
-            b_p = tf.get_variable('b_p', shape=(self.config.state_size, ), initializer=zero_init, dtype=tf.float64)
-            w = tf.get_variable('w', shape=(self.config.state_size, ), initializer=zero_init, dtype=tf.float64)
-            b = tf.get_variable('b', shape=(), initializer=zero_init, dtype=tf.float64)
+            W_q = tf.get_variable('W_q', shape=(self.config.state_size, self.config.state_size), initializer=xavier_init, dtype=tf.float32)
+            W_p = tf.get_variable('W_p', shape=(self.config.state_size, self.config.state_size), initializer=xavier_init, dtype=tf.float32)
+            W_r = tf.get_variable('W_r', shape=(self.config.state_size, self.config.state_size), initializer=xavier_init, dtype=tf.float32)
+            b_p = tf.get_variable('b_p', shape=(self.config.state_size, ), initializer=zero_init, dtype=tf.float32)
+            w = tf.get_variable('w', shape=(self.config.state_size, ), initializer=zero_init, dtype=tf.float32)
+            b = tf.get_variable('b', shape=(), initializer=zero_init, dtype=tf.float32)
 
         # ========================================================
         #  MatchLSTMCell class
@@ -849,7 +850,7 @@ class QASystemMatchLSTM(QASystem):
                                                      cell_bw=cell,
                                                      inputs=H_p,
                                                      sequence_length=self.context_length_placeholder,
-                                                     dtype=tf.float64)
+                                                     dtype=tf.float32)
             logging.debug('H_r_tuple is' + str(H_r_tuple))
 
         H_r = tf.concat(2, H_r_tuple)
@@ -865,11 +866,11 @@ class QASystemMatchLSTM(QASystem):
         max_context_length = self.config.max_context_length
       
         with tf.variable_scope('ansptr_LSTM'):
-            V = tf.get_variable('V', shape=(2*self.config.state_size, self.config.state_size), initializer=xavier_init, dtype=tf.float64)
-            W_a = tf.get_variable('W_a', shape=(self.config.state_size, self.config.state_size), initializer=xavier_init, dtype=tf.float64)
-            b_a = tf.get_variable('b_a', shape=(self.config.state_size, ), initializer=zero_init, dtype=tf.float64)
-            v =     tf.get_variable('v', shape=(self.config.state_size, ), initializer=zero_init, dtype=tf.float64)
-            c_v = tf.get_variable('c_v', shape=(), initializer=zero_init, dtype=tf.float64)
+            V = tf.get_variable('V', shape=(2*self.config.state_size, self.config.state_size), initializer=xavier_init, dtype=tf.float32)
+            W_a = tf.get_variable('W_a', shape=(self.config.state_size, self.config.state_size), initializer=xavier_init, dtype=tf.float32)
+            b_a = tf.get_variable('b_a', shape=(self.config.state_size, ), initializer=zero_init, dtype=tf.float32)
+            v =     tf.get_variable('v', shape=(self.config.state_size, ), initializer=zero_init, dtype=tf.float32)
+            c_v = tf.get_variable('c_v', shape=(), initializer=zero_init, dtype=tf.float32)
 
         # ========================================================
         #  AnswerPointerCell class
@@ -925,7 +926,7 @@ class QASystemMatchLSTM(QASystem):
             beta_log, _ = tf.nn.dynamic_rnn(cell=cell,
                                        inputs=input_dummy, # NOTE this input shouldn't matter? 
                                        sequence_length=seq_len_2, # only unroll twice
-                                       dtype=tf.float64)
+                                       dtype=tf.float32)
 
             logging.debug('beta_log' + str(beta_log))
             # NOTE we only need beta_0 and beta_1 in the intermediate steps in the cell
